+++ hypothesis-python/tests/django/toystore/test_given_forms.py
@@ -29,6 +29,7 @@ from tests.django.toystore.forms import (
     CustomerForm,
     DynamicForm,
     EmailFieldForm,
+    FileFieldsForm,
     InternetProtocolForm,
     ManyMultiValueForm,
     ManyNumericsForm,
@@ -134,6 +135,14 @@ class TestGetsBasicForms(TestCase):
         self.assertTrue(1 <= x.data["_decimal_one_to_five"] <= 5)
         self.assertTrue(1 <= x.data["_float_one_to_five"] <= 5)
         self.assertTrue(5 <= len(x.data["_string_five_to_ten"]) <= 10)
+
+    @given(from_form(FileFieldsForm))
+    def test_file_fields_form(self, x):
+        assert x.is_valid()
+        # form.data is empty, and form.files has one entry: file1
+        self.assertFalse(x.data)
+        self.assertTrue(set(x.files.keys()) == {"file1"})
+        self.assertTrue(x.files["file1"])
 
     @skipIf(not has_contrib_auth, "contrib.auth not installed")
     @given(from_form(UsernameForm))